refactor(docs): skeleton dx
