+++ apps/docs/content/components/skeleton/loaded.ts
@@ -1,36 +1,4 @@
-const App = `import {Card, Skeleton, Button} from "@nextui-org/react";
-
-export default function App() {
-  const [isLoaded, setIsLoaded] = React.useState(false);
-
-  const toggleLoad = () => {
-    setIsLoaded(!isLoaded);
-  };
-
-  return (
-    <div className="flex flex-col gap-3">
-      <Card className="w-[200px] space-y-5 p-4" radius="lg">
-        <Skeleton isLoaded={isLoaded} className="rounded-lg">
-          <div className="h-24 rounded-lg bg-secondary"></div>
-        </Skeleton>
-        <div className="space-y-3">
-          <Skeleton isLoaded={isLoaded} className="w-3/5 rounded-lg">
-            <div className="h-3 w-full rounded-lg bg-secondary"></div>
-          </Skeleton>
-          <Skeleton isLoaded={isLoaded} className="w-4/5 rounded-lg">
-            <div className="h-3 w-full rounded-lg bg-secondary-300"></div>
-          </Skeleton>
-          <Skeleton isLoaded={isLoaded} className="w-2/5 rounded-lg">
-            <div className="h-3 w-full rounded-lg bg-secondary-200"></div>
-          </Skeleton>
-        </div>
-      </Card>
-      <Button size="sm" variant="flat" color="secondary" onPress={toggleLoad}>
-        {isLoaded ? "Show" : "Hide"} Skeleton
-      </Button>
-    </div>
-  );
-}`;
+import App from "./loaded.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
